+++ CHANGELOG.md
@@ -8,7 +8,8 @@ All notable changes to **ValueStringBuilder** will be documented in this file. T
 
 ### Added
 
--  `LastIndexOf` to find the last occurence in the represented string
+- `LastIndexOf` to find the last occurence in the represented string.
+- `Replace<T>` added for generic replacement in the string builder.
 
 ## [0.9.5] - 2022-04-10